fix(honeycrisp): stop empty editor mount from clobbering note metadata
After logout and browser restart, notes reset to "Untitled / No content":
the first auto-selected note immediately, the rest when clicked. Body content
survives on the server, but the table-row metadata (title, preview, word
count) was overwritten with empty values permanently (issue #1590).

The editor persists metadata from every `docChanged` transaction. ySync's
initial render (`_forceRerender`) dispatches such a transaction synchronously
during `new EditorView(...)`, and it reports `docChanged` even when the
fragment is empty. On a signed-in relogin the body doc can render before its
WebSocket handshake delivers content, so that first dispatch (and any
pre-content sync frame) carried an empty document and wrote
`{ title: '', preview: '', wordCount: 0 }` to the row. Last-writer-wins made
the empty write durable. An explicit initial extraction after construction
made the same empty write a second time.

Skip persisting metadata when a ySync-origin transaction leaves the document
empty: that is the sync layer initializing or streaming in the body, not a
user edit. Real content arrives as a later non-empty sync transaction and
still persists; genuine user edits (including deliberately clearing a note)
are not sync-origin and still persist. Because the empty write never happens,
the row keeps the correct metadata the root doc already synced even if the
body never loads.

Move the pure derivation into `extract-metadata.ts` (`extractNoteMetadata`,
`isDocEmpty`) so the guard reads clearly and the logic is unit-testable.
